Refactor: EmergencyMap component
Refactored the EmergencyMap component to improve structure and readability.
